requirements.txt
python-telegram-bot[job-queue]==20.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
base58
telegram